+++ utilities/sourceEvents.py
@@ -136,6 +136,10 @@ def parse(content, gmaps):
         else:
             entry['category'] = eventTypeMap[entry['category']]
 
+        if pe['timeType'] == "ALL_DAY":
+            # skip all day event. (https://github.com/rokwire/events-manager/issues/1086)
+            continue
+            
         entry['originatingCalendarId'] = pe['originatingCalendarId']
         if pe.get("shareWithIllinoisMobileApp", "false") == "false":
             dataSourceEventId = pe.get("eventId", "")
@@ -265,17 +269,6 @@ def parse(content, gmaps):
             startTime = pe['startTime']
             startDateObj = datetime.strptime(startDate + ' ' + startTime + '', '%m/%d/%Y %I:%M %p')
             endDate = pe['endDate']
-            endDateObj = datetime.strptime(endDate + ' 11:59 pm', '%m/%d/%Y %I:%M %p')
-            # normalize event datetime to UTC
-            # TODO: current default time zone is CDT
-            entry['startDate'] = event_time_conversion.utctime(startDateObj, entry_location.get('latitude', 40.1153287), entry_location.get('longitude', -88.2280659))
-            entry['endDate'] = event_time_conversion.utctime(endDateObj, entry_location.get('latitude', 40.1153287), entry_location.get('longitude', -88.2280659))
-
-        if pe['timeType'] == "ALL_DAY":
-            entry['allDay'] = True
-            startDate = pe['startDate']
-            endDate = pe['endDate']
-            startDateObj = datetime.strptime(startDate + ' 12:00 am', '%m/%d/%Y %I:%M %p')
             endDateObj = datetime.strptime(endDate + ' 11:59 pm', '%m/%d/%Y %I:%M %p')
             # normalize event datetime to UTC
             # TODO: current default time zone is CDT